Added INT MIN edge case

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -8,7 +8,7 @@ int print_last_digit(int n)
 {
 	int last_digit;
 
-	last_digit = n > 0 ? n % 10 : -n % 10;
+	last_digit = n > 0 ? n % 10 : -(long int)n % 10;
 
 	_putchar(last_digit + '0');
 	return (last_digit);
